backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
mysql-connector-python==8.4.0
python-dotenv==1.0.1
pydantic==2.7.1
requests==2.32.3
python-jose[cryptography]==3.3.0
PyJWT==2.8.0